added validation of ingredients, moved logic to controllers
