+++ Server/constants.js
@@ -31,7 +31,8 @@ const OK = 200
  const SUCCESSFULLY_LOGGED_IN ='Successfully logged in'
  const ERROR_DURING_REGISTRATION = 'Error during registration'
  const MISSING_INGREDIENTS = 'Missing ingredients in request'
- const FAILED_TO_RETRIEVE_RECIPES ='Failed to retrieve recipes' 
+ const FAILED_TO_RETRIEVE_RECIPES ='Failed to retrieve recipes'
+ const INVALID_INGREDIENTS = "Invalid ingredients provided. Must be an array of strings with length between 2 and 20 characters long" 
 
 //Queries
  const FIND_ALL_USERS_BY_USERNAME = 'SELECT * FROM users WHERE username = $1'
@@ -39,6 +40,7 @@ const OK = 200
 
 
 module.exports = {
+  INVALID_INGREDIENTS,
   FAILED_TO_RETRIEVE_RECIPES,
   MISSING_INGREDIENTS,
     ERROR_DURING_REGISTRATION,
